Change generation of PyXB binding file to auto
The generation of the PyXB binding file has been made
to be initialised at the beginning of the program. It
is performed upon import of the schemadata module, before
accessing the binding file module itself. This is done to
alleviate unnecessary changes made by PyXB's own generation
method as it include system-specific information such as
XSDLocation. To suit this change, the argument -g has been
removed as it is now no longer needed. Include the latest
platform configuration spec 'platform_config.xsd' in the
folder 'schemaconfig'.