Move removal of temp files out of test class
When the removal of the temporary PyXB generated
binding file ./tmp/platformconfig.py occurs within
the related test class SchemadataTestCase, arc unit
fails as it refers to the file after deletion. Hence
the cleanup has been moved to occur on exit of the unittest.
However this means that arc unit does not run this cleanup
function.