Move binding gen code to separate module, add init
The code to generate the binding file is moved to the binding.py
module. An init() function has been added in the testApp and entry
point to create the necessary binding file before the start of the
tool or test. This init function must be called before importing
any modules that use the binding file 'platformconfig.py'